tour + refactoring
