+++ src/components/Modals/MultiplePlans/TextInputModal.vue
@@ -50,7 +50,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '@/components/Modals/PlanModalDropdown.scss';
+@import '@/components/Modals/MultiplePlans/PlanModalDropdown.scss';
 
 .textInput {
   &-label {
